# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  dtos.ErrResp:
    properties:
      error:
        type: string
      path:
        type: string
      timestamp:
        type: string
    type: object
...
  dtos.LevelCreateDTO:
    properties:
      img_uri:
        type: string
      name:
        maxLength: 20
        type: string
    required:
    - img_uri
    - name
    type: object
  dtos.LevelReadDTO:
    properties:
      id:
        type: integer
      img_uri:
        type: string
      name:
        type: string
    type: object
  dtos.LevelUpdateDTO:
    properties:
      id:
        type: integer
      img_uri:
        type: string
      name:
        maxLength: 20
        type: string
    required:
    - id
    - img_uri
    - name
    type: object
...
  dtos.TopLevelsDTO:
    properties:
      count_of_plays:
        type: integer
      id:
        type: integer
      img_uri:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This site documents the endpoints for the smaash web app, allowing
    easy testing
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: SMAASH API documentation
  version: "1.0"
paths:
...
  /levels:
    get:
      consumes:
      - application/json
      description: Reads all levels
      produces:
      - application/json
      responses:
        "200":
          description: returns all levels
          schema:
            items:
              $ref: '#/definitions/dtos.LevelReadDTO'
            type: array
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
    post:
      consumes:
      - application/json
      description: Creates a new level
      parameters:
      - description: dto for creating a new level
        in: body
        name: level_create_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.LevelCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: returns newly created level
          schema:
            $ref: '#/definitions/dtos.LevelReadDTO'
        "400":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
  /levels/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a level with the given id
      parameters:
      - description: id of desired level
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: doesn't return anything
          schema:
            type: ""
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
    get:
      consumes:
      - application/json
      description: Reads a level by it's id
      parameters:
      - description: the id of the desired level
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns the desired level
          schema:
            $ref: '#/definitions/dtos.LevelReadDTO'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
    put:
      consumes:
      - application/json
      description: Updates the level with the given id
      parameters:
      - description: dto for updating a level
        in: body
        name: level_update_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.LevelUpdateDTO'
      - description: id of desired level
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: doesn't return anything
          schema:
            type: ""
        "400":
          description: id from url and id from request body doesn't match
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
  /levels/{id}/img:
    get:
      consumes:
      - application/json
      description: Returns an uploaded level image
      parameters:
      - description: id of desired level
        in: path
        name: id
        required: true
        type: integer
      produces:
      - multipart/form-data
      responses:
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image for level with given id
      parameters:
      - description: id of desired level
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: returns newly created image's URI
          schema:
            type: string
        "400":
          description: no file sent
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "415":
          description: invalid media type
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - levels
...
  /stats/top/levels:
    get:
      consumes:
      - application/json
      description: Reads the most played levels
      produces:
      - application/json
      responses:
        "200":
          description: returns levels ordered by count of matches played
          schema:
            items:
              $ref: '#/definitions/dtos.TopLevelsDTO'
            type: array
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - stats
...
swagger: "2.0"